+++ geotrek/signage/tests/test_command_signage.py
@@ -65,11 +65,16 @@ class SignageCommandTest(TestCase):
         structure = StructureFactory.create(name='structure')
         filename = os.path.join(os.path.dirname(__file__), 'data', 'signage.shp')
         call_command('loadsignage', filename, type_field='label', name_field='name',
-                     condition_field='condition', structure_default='structure',
+                     condition_field='condition', manager_field='manager', sealing_field='sealing',
+                     structure_default='structure',
                      description_field='descriptio', year_field='year', code_field='name', verbosity=1, stdout=output)
         self.assertIn('Signages will be linked to %s' % structure, output.getvalue())
         self.assertIn("SignageType 'type' created", output.getvalue())
         self.assertIn("Condition Type 'condition' created", output.getvalue())
+        self.assertIn("Organism 'manager' created", output.getvalue())
+        self.assertIn("Organism 'manager2' created", output.getvalue())
+        self.assertIn("Sealing 'sealing' created", output.getvalue())
+        self.assertIn("Sealing 'sealing2' created", output.getvalue())
         value = Signage.objects.all()
         names = [val.name for val in value]
         years = [val.implantation_year for val in value]
@@ -108,6 +113,10 @@ class SignageCommandTest(TestCase):
         call_command('loadsignage', filename, type_default='label', name_field='name',
                      condition_field='wrong_condition_field', stdout=output)
         call_command('loadsignage', filename, type_default='label', name_field='name',
+                     manager_field='wrong_manager_field', stdout=output)
+        call_command('loadsignage', filename, type_default='label', name_field='name',
+                     sealing_field='wrong_sealing_field', stdout=output)
+        call_command('loadsignage', filename, type_default='label', name_field='name',
                      description_field='wrong_description_field', stdout=output)
         call_command('loadsignage', filename, type_default='label', name_field='name',
                      year_field='wrong_implantation_year_field', stdout=output)
@@ -119,9 +128,9 @@ class SignageCommandTest(TestCase):
                      eid_field='wrong_eid_field', stdout=output)
         elements_to_check = ['wrong_type_field', 'wrong_name_field', 'wrong_condition_field', 'wrong_code_field',
                              'wrong_description_field', 'wrong_implantation_year_field', 'wrong_structure_field',
-                             'wrong_eid_field']
+                             'wrong_eid_field', 'wrong_sealing_field', 'wrong_manager_field']
         self.assertEqual(output.getvalue().count("set a default value"), 2)
-        self.assertEqual(output.getvalue().count("Change your"), 6)
+        self.assertEqual(output.getvalue().count("Change your"), 8)
         for element in elements_to_check:
             self.assertIn("Field '{}' not found in data source".format(element),
                           output.getvalue())